requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-ai-formrecognizer
azure.identity
azure-core
openai
requests
azure-storage-blob
azure-search-documents
PyMuPDF
python-docx
pyodbc
tiktoken
langchain
